refactor: rename classic/next mods to ui_classic/ui_next
- bare `mod next` at the top of a file reads ambiguously; the `ui_`
  prefix makes the backend split obvious at a glance
- covers `interactive/`, `favorites/`, and `ui.rs`
